Faster os.walk() replacement on Linux
Take advantage of the fact that readdir_r() returns the file type with each
directory entry.

Non-scientific data point: this reduces an "attic list" from ~500ms to ~240ms
for a 87G repository on a USB NTFS drive that consists of ~17k individual
segment files.  An "attic extract" for a small file is similarly about 300ms
faster (2.6 -> 2.3 sec).